Factor the UPC API into a separate class
