+++ app/src/main/java/com/example/ezvault/EditItemDetails.java
@@ -50,6 +50,9 @@ import com.google.android.gms.tasks.Task;
 import com.google.android.gms.tasks.Tasks;
 import com.google.android.material.textfield.TextInputLayout;
 import com.google.firebase.Timestamp;
+import com.google.mlkit.vision.codescanner.GmsBarcodeScanner;
+import com.google.mlkit.vision.codescanner.GmsBarcodeScannerOptions;
+import com.google.mlkit.vision.codescanner.GmsBarcodeScanning;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -232,6 +235,7 @@ public class EditItemDetails extends Fragment {
             dialog.show();
         });
 
+
         serialText.setOnClickListener(v -> serialText.showDropDown());
         serialText.setOnFocusChangeListener((v,f) -> serialText.showDropDown());
         serialAdapter = new ArrayAdapter<>(requireContext(), android.R.layout.select_dialog_singlechoice);
@@ -262,6 +266,9 @@ public class EditItemDetails extends Fragment {
                 });
             }
         });
+
+        TextInputLayout descriptionLayout = view.findViewById(R.id.edit_details_desc_layout);
+        descriptionLayout.setEndIconOnClickListener(barcodeListener);
 
 
         //update to new values on save
@@ -345,5 +352,21 @@ public class EditItemDetails extends Fragment {
         saveButton.setClickable(canInteract);
         saveButton.getBackground().setAlpha(canInteract ? 255 : 112);
     }
+
+    protected View.OnClickListener barcodeListener = new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            GmsBarcodeScannerOptions options = new GmsBarcodeScannerOptions.Builder()
+                    .enableAutoZoom()
+                    .build();
+            GmsBarcodeScanner scanner = GmsBarcodeScanning.getClient(EditItemDetails.this.getActivity(), options);
+            scanner.startScan().addOnSuccessListener(
+                    barcode -> {
+                        upcAPI api = new upcAPI();
+                        api.upcLookup(barcode.getRawValue(), getView().findViewById(R.id.edittext_item_description), getActivity());
+                    }
+            );
+        }
+    };
 
 }
